Point the User Guide roadmap sections at a new duplicate of the roadmap in the Developer Guide.  It's confusing to keep around different versions of a 'current' roadmap, and directing people to an online version seems like a sensible idea.
Note that 1.0.x apparently did not have a roadmap.

(I also considered choosing one version of the user guide to be the
'canonical' one, and forwarding the others.  However, the obvious location
for a canonical version is either 'latest stable' or 'unstable', both of
which would be problematic.  This seemed easier, and has the advantage of
putting the roadmap next to the people doing the work.)