Remove sub-second formatting and parsing for Offset.
Specifically, remove the 'f' standard pattern, and the 'f' and 'F' (and '.')
custom patterns.  Adjust all tests to remove expectations that include
fractional seconds, and remove any (most) that are now redundant.

Note that the documentation changes here anticipate adjusting the resolution
of Offset to seconds (coming in this branch) and also making it an error to
use unrecognised format characters (coming in issue 308).

(The tests do not currently all pass, but I believe that the only failures
are due to Offset.{MinValue,MaxValue} having fractional seconds, so they'll
be fixed automatically when Offset moves to seconds.)